Add ASN.1 object support
With this change, PDF objects, which can contain a DER-encoded
ASN.1 objects inside them, now will also create a tree. This tree
represents the parsed ASN.1 object and behaves the same way as the PDF
tree.

ASN.1 parsing is done by Bouncy Castle. Handling of those library
primitives is done via the Asn1TreeNode family of classes.

Icons are composed from the same icon set, that was used before.

Insertion of the ASN.1 object subtree into the big PDF tree is done
inside TreeNodeFactory. As of now, the following objects are expanded
into ASN.1 subtrees:

1. /Contents string inside a /Sig or /DocTimeStamp dictionary.
2. CRL, certificate, OCSP and timestamp streams inside a /DSS
   dictionary.
3. Certificates inside a seed value dictionary.

I have not found examples of (3) and a timestamp inside /DSS, so they
would still need some real-world testing.

As of this comment, there are no convenient display names for ASN.1
objects. Also, since this has no access to the specs, IMPLICIT ASN.1
tags can be parsed incorrectly (i.e. show a different underlying type
that what is expected in a spec or be parsed as EXPLICIT). This will
be handled in a later commit.