Add object identifier "deciphering"
With this commit ASN.1 nodes, which show OBJECT IDENTIFIER types,
will display both the numeric value and its somewhat reference
descriptive name.

Mapping is done inside the OidReference class with all the data. Ideally
I wanted to just use Bouncy Castle, as it would be weird to not have
such functionality there. But, sadly, there does not seem to be a way to
just "take OID and return a string". They do not seem to have a
"global" map, only smaller maps within specific classes (like digest,
DS, ...) and they are missing some of the less crypto-related stuff.

So OidReference was born with the data from RFCs and the wider Internet.
It seems like I've got the most common stuff covered, but it should be
easy do add mappings later, if we've missed something.