Add field naming and fix tagging for X.509 certificates
With this commit now most X.509 certificates inside the PDF context
now should have properly named fields instead of just having generic
ASN.1 type names.

This is handled via the Corrector family of classes. They take a ASN.1
subtree and improve the field nodes as it goes through its validation.

Some of the things a corrector can do:

1. Replace generic node names with the ones from the spec.
2. Fix UNIVERSAL types, which are hidden behind IMPLICIT tags. This
   cannot be done without spec information.
3. Provide descriptive names for some of the values in the object. Like
   INTEGER or ENUMERATED known values, or bit flags inside a BIT STRING.
4. Expand DER-encoded objects, which are stored as an OCTET STRING. An
   example of this is X.509 certificate extensions.

You can get more information in the Corrector javadoc.

In general, they behave like validators, so there is a lot of code,
which is relatively simple. But there is _a lot_ of it.

While I was implementing this, I've found, that Bouncy Castle had
validation for most of these types. For example, a SignedData class,
which takes an encodable object and makes it into a Java comparable one.
But after checking further, it seemed that it would be easier and more
featureful to do it in RUPS for the following reasons:

1. You would need to map the Java object fields back to the ASN.1 object
   ones to fix the tree nodes. So you would need to know the spec anyway
   to do this. And at this point you might as well just check it
   yourself, as it is not difficult.
2. There classes, for obvious reasons, have an all-or-nothing approach.
   If any field is incorrect, it will just throw an exception, and the
   whole object won't have any improvements. Correctors on the other
   hand handle this in a mroe graceful manner. For example, if a version
   field is an ENUMERATED instead of an INTEGER for some reason, then
   only that field won't get a name with the Corrector, and everything
   else will still look correct. This should be easier for visual
   inspection.

Overall, it seems to be working pretty good, but since some of the types
are massive, there is quite a lot of code to support them.