Add saving of a textual representation of ASN.1 trees
Before "Save to File" and "Save Raw Bytes to File" were the same for
ASN.1 objects. With this change "Save to File" will save a JSON
textual representation of the ASN.1 tree.

The result is the representation of the _tree_, not of the ASN.1 object
itself. I.e. it will have additional data, like having children under
bit flag strings, expanding data under DER-encoded octet strings, all
values are returned as strings, etc.