Fix selection jumping, when walking through ASN.1 trees
I am not 100% sure, how this happens, but PdfTreeExpansionListener can
receive somewhat odd paths. If you try to create a path from the last
path component to the root via node parents, it will not match the path,
which is passed to the listener.

This can happen, for example, if you open a signature under /AcroForm,
but then try to walk through it under page /Annots. In this case, if you
expand a node with only one child, it will cause the view to jump back
to the /AcroForm tree.

Change the code to expand the received path with the added child instead
of building a new one from child. I have a feeling it was written
differently before for a reason and some bugs may pop-up, but I haven't
noticed anything broken after the change.