Improve "Open File" UX
Now there is a "Most Recently Used" list of opened files, which is
available in the file menu. It also makes it so that when you use the
"Open File" dialog, it used the directory of the latest open file.

MRU list stores 20 entries and is handled by MruListHandler within
RupsConfiguration. Persistent storage is done via Preferences (ex. it
will be stored in registry on Windows).

Also now, since MRU list could contain files, which were already
deleted, there is now a better message, if the file you are trying to
open is missing.

OPEN_FILE_EVENT was removed in the continuing effort to remove Observer.
Now there is a listener list in the controller for such an event with
the Consumer<File> interface.

There were somewhat major changes in RupsMenuBar:
* "items" hash map was removed. It was used only to disable/enable items
  during tab change events. But there is no reason to have a hash map,
  when we could just store references to the affected menu items
  directly.
* Removed storage of most of the actions, as they were not used
  afterwards.
* Added some addItem Language overloads to make it a bit cleaner.
* Formatting changes.