Remove deprecated Observer and Observable usage
With this the RupsEvent and its package was removed. Usages were
replaced with the following:

1. Somewhat global events (like document opened/closed) are now handled
   by IRupsEventListener. Implementers of this interface handle the
   global events and may forward them further in any way they like.

2. Events from PdfObjectPanel were extracted to
   IPdfObjectPanelEventListener interface and now only
   PdfReaderController explicitly handles them. Not quite sure, why they
   were on a somewhat global level originally.

3. Events, like _requests_ (not the aftermath) of stuff, like open/close
   were replaced with direct calls or individual functional listener
   interfaces.

FileCloseAction was inlined into a lambda. There is no real need to have
a full class, when the handler just calls one methods on an object.

FileCompareAction was removed, since it is currently unused and we would
want to reuse PdfFileOpenAction for it in the future, so that MRU also
works there.