Improve .exe JRE compatibility handling
RUPS now requires at least Java 11, but we still had minVersion set to
1.7. This caused there to be a very non-descriptive error message, when
you try launch the executable with something like Java 1.8:

"A Java Exception has occurred"

Version was bumped to 11, so now there is a proper message, that tells
you to install JRE 11.

In the older versions of this plugin, in such cases it would open a
browser with the Oracle Java 1.8 download page. In the updated one it
doesn't open anything. To make it more user friendly, as it is in no way
obvious as to where you can get just the JRE, I've added a download link
for the Eclipse Temurin JRE.

Also, before launch4j would not work properly with minVersion 11, as it
just checked the registry keys, which are not set by default in the
newer releases of JRE. In the current versions it checks JAVA_HOME and
PATH first, so now it works properly. But without setting 'path' path
explicitly and just using the default, the plugin produces a warning, so
it was added explicitly.