Fix macOS application bundle build
The plugin we were using for building bundles was not updated in a
while, so it doesn't work with Java 11. It also doesn't support ARM
properly.

Changed it to use a fork. This is turn removed one of the vulnerable
dependencies.

Now it works properly with Java 11 and the loader inside is a Universal
binaries, so it works properly on ARM.

Would be great to sign it, so that we could release it...