Add PDF content stream parsing logic
Initially the idea was to just use PdfTokenizer from iText, but it had
some problems, like not preserving token positions in the input data,
skipping whitespace outright, have hard errors for invalid output, etc.
For more example see PdfContentStreamParser docs.

This will be used as a reference model for the editor. Ideally we would
have the same model in the editor to limit memory consumption, but
having a separate one allows us some flexibility, which might be useful
when implementing static analysis.