Make RUPS open document in read-only mode by default
Since the iText library upgrade we have an issue, that some of the
information is missing from the trailer. For example, the /Info
dictionary is always missing, if the PDF document is opened in the
read-write mode. But even before that, there were some modifications
happening to the data, like the modification date changing.

As a workaround for now we will start opening all the files in the
read-only mode by default. In this case it should show data much closer
to what is in the original document.

If you want to edit the document, you can use the "Reopen as Owner"
option, which reuses the code added for user/owner password protection
handling.

This still leave the issue, that in the read-write mode the data doesn't
match, but it would take more effort to solve...