Bring back file URI support in open dialog
Before some of my older refactoring changes, FilePathPreProcessor was
used to allow to open "file:" URIs and to also decode URL-encoded file
names. But I've removed it in the Open dialog, as it was pretty hacky
and could result in not opening files in some cases.

So I've made some changes to the FilePathPreProcessor:
* If a file exists on disk at the specified path, no conversion happens.
* URL-decoding of paths only happens, if no file at the original path
  exists and there is a file at the decoded path.
* File URI handling is now done via the URI and File classes, which
  should make it more robust.

With that in mind, there was still a problem with the Open dialog.
JFileChooser only returns a File, not the path provided in the dialog.
And there doesn't seem to be a way to get it without UI class hackery.
In the original implementation it searched for "file:" within the path
and discarded everything before it. This comes from the fact, that if,
for example, you put "file:///C:/3142b600.pdf" in the dialog, you would
get something like this from the File API:
"C:\selected_dir\file:\C:\3142b600.pdf"

To handle this in a less hacky way, there is an EnhancedFileSystemView
now, which is passed to the JFileChooser. It just wraps the original
FileSystemView, but in the File creation methods, which involve String
file names or paths (createFileObject, getChild), they now go through
the FilePathPreProcessor. This way all the path transformation will
happen before File conversion, which is much cleaner.