Fix some SonarQube issues, which cause logic change
In RupsController the dimensions are now copied in the constructor.
This shouldn't change anything in practice, as, from what I can tell,
if screen dimensions change, it should just give you a different
dimensions object instead of modifying the existing one.

Try-with-resources is used now during save, which changes the error
handling logic a bit, but not in a way, that really matters.

3 pixel offset in ContextMenuMouseListener pop-up was removed to make
it in line with all the other context menus we have in the app.

In DictionaryTableModel oldName init was moved closer to usage, as
getCorrectKey doesn't affect the keys array.

In StreamTrackingPdfCanvasProcessor Stack was replaced with Deque,
shouldn't change anything in practice there.

".trim().length() == 0" string checks were replaced with ".isBlank()",
which is _not_ 100% the same, but should be fine here.

There are some minor changes in the SaveTo* dialog code, but the only
one, that should be visible to the user is that now a failure to save
will also be shown in a Toast at the bottom.

As for the encoding changes, they are not that obvious... Before it was
using the default encoding, which, most likely, would have been UTF-8,
at least on *nix. For places, like console logs, where it doesn't really
matter, I've changed encoding explicitly to UTF-8. But in places, when
using text was already a hack (like loading and saving binary PDF
streams), I've set encoding explicitly to Latin-1. Don't think, that
will change a lot in practice. Hopefully I'll be able to fix that
properly with the new content stream editor.