Handle exception thrown in ObjectLoader::doInBackground
As it turns out, SwingWorker still called done, even if the body of the
task itself threw an exception. So it just pushes the open event, even
though `nodes` is null.

What's annoying, is that there doesn't seem to be a good mechanism for
getting that exception... You either need to handle it yourself within
`doInBackground` and then read that exception data in `done`, or abuse
`SwingWorker::get` and catch an `ExecutionException`.

See issue #36 for discovery.