Add ability to manipulate stream filters
When you open a document as owner, you can now manipulate stream
filters via the context menu. The available options at the moment are
the following:

1. Remove all filters from a stream. I.e. save the decoded stream into
   the file.
2. Add an `ASCII85Decode` filter.
3. Add an `ASCIIHexDecode` filter.
4. Add a `BrotliDecode` filter with the default compression level.
5. Add a `FlateDecode` filter with the default compression level.
6. Add a `RunLengthDecode` filter.

When you add a filter, they are layered on top, not replaced. So you
could test some of the more unorthodox configurations.

Additionally, when you save a stream now, it is saved without any
filters applied by default. And now the `Length` value is correctly
updated on the stream.