Add corrector for AuthenticatedData
There is a lot of overlap with SignedData, so some of the common types
were extracted into separate corrector classes. These correctors are
not mapped to OIDs, so they are put into a separate `types` package.

We've already had such correctors, so, in the future, I'll move them
around, so that non-mapped correctors are always in a `types`
sub-package.